Replace repeated function name rpn with internal function.
This was an idea that Dane and I had kicked around for simplifying a
function with multiple pattern matches. Instead of this:

        foobar(a) -> 1;
        foobar(b) -> 2;
        ...
        foobar(z) -> 26.

we replace it with something that has only one foobar:

        foobar(Arg) ->
            F = fun(a) -> 1;
                   (b) -> 2;
                   ...
                   (z) - 26
                end,
            F(Arg).

I think this contributes greatly to improved readability, especially if
the outer function name is long and descriptive.